Renamed git.js to git-helper.js to prevent errors in Windows
