+++ git-helper.js
